requirements.txt
Django == 2.1.11
django-cors-headers == 2.4.0
djangorestframework == 3.9.2
HttpRunner == 1.5.15
PyMySQL == 0.9.3
beautifulsoup4 == 4.6.3
tornado == 5.1.1
django-celery == 3.2.2
pillow
djangorestframework-jwt
xlrd
xlsxwriter
django-crispy-forms
django-formtools
django-reversion
django-import-export
httplib2
future
django-rest-swagger
